Control.Foldl.NonEmpty

Description

This module provides a Fold1 type that is a "non-empty" analog of the Fold type, meaning that it requires at least one input element in order to produce a result

This module does not provide all of the same utilities as the Control.Foldl module. Instead, this module only provides the utilities which can make use of the non-empty input guarantee (e.g. head). For all other utilities you can convert them from the equivalent Fold using fromFold.

Documentation

data Fold1 a b

A Fold1 is like a Fold except that it consumes at least one input element

Constructors

Fold1 (a -> Fold a b)

fold1 :: Foldable1 f => Fold1 a b -> f a -> b

Apply a strict left Fold1 to a NonEmpty list

fromFold :: Fold a b -> Fold1 a b

Promote any Fold to an equivalent Fold1

toFold :: Fold1 a b -> Fold a (Maybe b)

Promote any Fold1 to an equivalent Fold

nonEmpty :: Fold1 a (NonEmpty a)

Fold all values within a non-empty container into a NonEmpty list

sconcat :: Semigroup a => Fold1 a a

Fold all values within a non-empty container using (<>)

head :: Fold1 a a

Get the first element of a non-empty container

last :: Fold1 a a

Get the last element of a non-empty container

maximum :: Ord a => Fold1 a a

Computes the maximum element

maximumBy :: (a -> a -> Ordering) -> Fold1 a a

Computes the maximum element with respect to the given comparison function

minimum :: Ord a => Fold1 a a

Computes the minimum element

minimumBy :: (a -> a -> Ordering) -> Fold1 a a

Computes the minimum element with respect to the given comparison function